Makes UMD primary and ES available as module.js
