+++ shared/websocket/module.js
@@ -1,0 +1,1 @@
+export { default } from './src/createWebsocket.js';
